Fix some compile warnings on MSVC, reorganize some build files

+++ src/builtin_functions/file_changed.cpp
@@ -20,7 +20,7 @@ namespace file_changed_function {
         if (modifiedTime != prevModified
                 || filename != prevFilename) {
             prevFilename = filename;
-            prevModified = modifiedTime;
+            prevModified = (int) modifiedTime;
             return true;
         } else {
             return false;
